% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.semiparMF}
\alias{plot.semiparMF}
\title{Plot Convergence History}
\usage{
\method{plot}{semiparMF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{semiparMF}.}

\item{...}{Additional graphical parameters.}
}
\description{
Plots the Mean Squared Prediction Error (MSPE) across iterations.
}
