% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.semiparMF}
\alias{predict.semiparMF}
\title{Predict Method for SemiparMF}
\usage{
\method{predict}{semiparMF}(object, new_high_freq, new_z, new_w, ...)
}
\arguments{
\item{object}{An object of class \code{semiparMF}.}

\item{new_high_freq}{Numeric array (N x T_new x K) for the high-frequency covariate.}

\item{new_z}{Numeric matrix (N x T_new) for the parametric covariate.}

\item{new_w}{Numeric matrix (N x T_new) for the neighborhood covariate.}

\item{...}{Additional arguments.}
}
\value{
A matrix (N x T_new) of predicted values.
}
\description{
Generates predictions for new data.
Reference: "Predicted values are calculated by simply adding up the scores... and the linear combination".
}
