% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.semiparMF}
\alias{print.summary.semiparMF}
\title{Print Summary for SemiparMF}
\usage{
\method{print}{summary.semiparMF}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.semiparMF}.}

\item{...}{Additional arguments.}
}
\value{
The input object \code{x} is returned invisibly.
}
\description{
Print Summary for SemiparMF
}
