% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.semiparMF}
\alias{summary.semiparMF}
\title{Summary Method for SemiparMF}
\usage{
\method{summary}{semiparMF}(object, ...)
}
\arguments{
\item{object}{An object of class \code{semiparMF}.}

\item{...}{Additional arguments.}
}
\description{
Summary Method for SemiparMF
}
